requirements.txt
Pillow==11.1.0
toml==0.10.2
